% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabifyr.R
\docType{package}
\name{syllabifyr}
\alias{syllabifyr}
\alias{syllabifyr-package}
\title{Syllabify: A package for doing tidy syllabification}
\description{
This is a package to do tidy syllabification of phonetic transcriptions.
}
